+++ CHANGES.md
@@ -1,3 +1,7 @@
+## ?
+### Changed
+- `Conformist.Field.optional` is now `Conformist.Field.is_optional` to avoid confusion with `Conformist.optional`
+
 ## 0.5.0 - 2021-04-12
 ### Added
 - Conformist type `datetime` that decodes to `Ptime.t`. This replaces `date` which has been deprecated.